Rename table_block.fields to table_block.blocks
